refactor JwtService
